+++ management/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.39.0 (Unreleased)
+
+* **Enhancement** Added the `NetworkPolicyServer` property to the `GatewayTypeRADIUS` data model. [#336](https://github.com/patrickcping/pingone-go-sdk-v2/pull/336)
+
 # v0.38.0 (2024-03-15)
 
 * **Enhancement** Added the `DefaultTargetUrl` property to the `ApplicationSAML` data model. [#325](https://github.com/patrickcping/pingone-go-sdk-v2/pull/325)